✨ feat: add cmdk package and implement Command component for enhanced command palette functionality
